[ soundness check ] Introduce a soundness check.
SoC designs /are/ directed graphs.
In fact they are k-partite graphs if we view each module's ports as a partition of a graph.
The wiring between ports represents the physical wires.

Our soundness check (not a proof but it is a proof) states that well-typed modules are sound /linearly wired/ iff each port is used once.
We check this by interpreting our model to a graph and check the degrees of each leaf node representing a port is one.
Fanouts are translated to a node in that accepts a single input and branches for each output.
Multiplexers are similarly translated but with an additional input for the control wire.